fixup! Move Guava's Iterators and Iterables into lib module
